Minio: Support creating multiple buckets at install
Alter the behavior of `defaultBucket` and the post-install hook for creating buckets at installation to support the creation of multiple buckets.

Moved from `defaultBucket` to `defaultBuckets`, which now provides an array of items with `name`, `policy`, and `purge` properties. Name is required, policy defaults to `none`, and purge requires a hard set of `true` to cause removal of existing data if present.

[ci skip]