Minio: Move bucket creation logic to script in ConfigMap
Move the scripting for creating the buckets from using a mangled `command:` entry into a script (`initialize`) that is included in the ConfigMap for the chart. This allows us better maintainability do to the way that `sh -c 'whargarbl; #comment ; missed code'` is handled, as well as allowing expanded scripting.

Also moved the job and file name to `buckets` from `bucket`