NGINX/Minio: add Minio to NGINX Ingress
Add matching template functionality to that of GitLab and Registry for Minio. I've copied all existing templates, with a change of name to reflect the focus on `minio`.

I've rearranged a few of the existing functions for a consistent format of function, chart order. This is always gitlab, registry, minio.